Update editAssets.py
Now takes a tsv/csv as input and automatically converts it into JSON data.
It's now game agnostic.
Special edits to files are no longer hardcoded but instead contained in a separate folder.

+++ editAssets.py
@@ -12,7 +12,7 @@ from UnityPy.enums.TextureFormat import TextureFormat
 STRUCTS_PATH = ""
 TYPES = [
     # Images
-    "Sprite",
+    # "Sprite",
     "Texture2D",
     # Text (filish)
     # "TextAsset",
@@ -26,36 +26,34 @@ TYPES = [
 STRUCTS = {}
 MB_JSON = {}
 T2D_JSON = {}
-DEBUG_DICT = {}
+SPECIAL_OVR = {}
 tlDebug = False
-src = ""
-
-
+dataFolder = ""
 # Credit to K0lb3 for the code I based this off of, as well as the unityPy library in general:
 # https://github.com/K0lb3/Romancing-SaGa-Re-univerSe-asset-downloader/blob/24388c660c3ed40665d840b6cf46bbbd48d63b8e/lib/asset.py
-
-
 def main():
-    holoErrorDirectory = os.getcwd()
-    while not (os.path.isfile((os.path.join(holoErrorDirectory, "hololiveERROR.exe")))):
-        if os.path.dirname(holoErrorDirectory) == holoErrorDirectory:
-            input("hololiveERROR.exe not found!")
+    rootDirectory = os.getcwd()
+    while not (os.path.isfile((os.path.join(rootDirectory, "GameAssembly.dll")))):
+        if os.path.dirname(rootDirectory) == rootDirectory:
+            input("GameAssembly.dll could not be found in any above directory!")
             return 1
         else:
-            holoErrorDirectory = os.path.dirname(holoErrorDirectory)
+            rootDirectory = os.path.dirname(rootDirectory)
             # Move up the directory tree, one directory at a time
-
-    TL_path = os.path.join(holoErrorDirectory, "TranslationPatch")
+    TL_path = os.path.join(rootDirectory, "TranslationPatch")
     if not (os.path.isdir(TL_path)):
         input("TranslationPatch folder not found!")
         return 1
-    global src
-    src = os.path.join(holoErrorDirectory, "hololiveERROR_Data")
+    global dataFolder
+    for dirName in os.listdir(rootDirectory):
+        if os.path.isdir(os.path.join(rootDirectory, dirName)):
+            if "_Data" in dirName:
+                dataFolder = os.path.join(rootDirectory, dirName)
+                break  # We've got our _Data directory
     # print(src)
-    if not (os.path.isdir(src)):
-        input("hololiveERROR_Data folder not found!")
+    if not (os.path.isdir(dataFolder)):
+        input("Data folder not found!")
         return
-
     if os.path.isfile(os.path.join(TL_path, "tlDebug.txt")):
         print("Would you like the subtitle IDs to be printed alongside the text?")
         answerTemp = input("y/n\n")
@@ -68,17 +66,28 @@ def main():
             tlDebug = True
         else:
             tlDebug = False
-
     subFolders = [f.path for f in os.scandir(TL_path) if (f.is_dir() and f.path[f.path.rfind("\\") + 1::] != "exeData"
-                                                          and f.path[f.path.rfind("\\") + 1::] != "debugOverrides")]
+                                                          and f.path[f.path.rfind("\\") + 1::] != "specialOverrides")]
     i = 1
     for sf in subFolders:
         print(str(i) + ") " + sf[sf.rfind("\\") + 1::])
         i = i + 1
-
-    fileList = ["level0", "sharedassets0.assets"]  # These are the two files that require editing.
-    if os.path.isfile(os.path.join(src, fileList[0] + "_JP")):
+    if os.path.isfile(os.path.join(dataFolder, "sharedassets0.assets_JP")):
         print(str(i) + ") Restore original language")
+
+    fileList = []  # These are the files that require editing.
+    i = 0
+    while os.path.isfile(os.path.join(dataFolder, "level" + str(i))):
+        fileList.append("level" + str(i))
+        i = i + 1
+    i = 0
+    while os.path.isfile(os.path.join(dataFolder, "sharedassets" + str(i) + ".assets")):
+        fileList.append("sharedassets" + str(i) + ".assets")
+        i = i + 1
+    # if os.path.isfile(os.path.join(dataFolder, "globalgamemanagers.assets")):
+    #    fileList.append("globalgamemanagers.assets")
+    if os.path.isfile(os.path.join(dataFolder, "resources.assets")):
+        fileList.append("resources.assets")
 
     TL_num = -1
     while not (0 <= TL_num - 1 <= len(subFolders)):
@@ -87,23 +96,23 @@ def main():
         except:
             print("Not a valid number.")
 
-    global DEBUG_DICT
+    global SPECIAL_OVR
     global T2D_JSON
 
-    if os.path.isdir(os.path.join(TL_path, "debugOverrides")):
-        print("debugOverrides found, patching files manually")
-        debugFileList = [f.path for f in os.scandir(os.path.join(TL_path, "debugOverrides")) if
+    if os.path.isdir(os.path.join(TL_path, "specialOverrides")):
+        debugFileList = [f.path for f in os.scandir(os.path.join(TL_path, "specialOverrides")) if
                          (f.is_file() and (f.name.endswith(".json")))]
         for f in debugFileList:
-            DEBUG_DICT[f[f.rfind("#") + 1:f.rfind("."):]] = f
+            SPECIAL_OVR[f[f.rfind("#") + 1:f.rfind("."):]] = f
 
     TL_num = TL_num - 1
     if TL_num == len(subFolders):
         # User wants to restore japanese language. Do that and exit.
         for resFile in fileList:
-            curFile = os.path.join(src, resFile)
-            if os.path.isfile(curFile): os.remove(curFile)
-            os.rename(curFile + "_JP", curFile)
+            curFile = os.path.join(dataFolder, resFile)
+            if os.path.isfile(curFile + "_JP"):  # Only do this is there's a corresponding JP version.
+                if os.path.isfile(curFile): os.remove(curFile)
+                os.rename(curFile + "_JP", curFile)
         return
     TL_path = subFolders[TL_num]
 
@@ -119,29 +128,23 @@ def main():
     if not (os.path.isfile(STRUCTS_PATH)):
         input("Error, file is not alongside typetrees.json, did you change the working directory?")
         return 1
-
     for resFile in fileList:
-        curFile = os.path.join(src, resFile)
+        curFile = os.path.join(dataFolder, resFile)
         if not os.path.isfile(curFile + "_JP"):  # If JP file doesn't exist, move the original one.
             if os.path.isfile(curFile):
                 os.rename(curFile, curFile + "_JP")
             else:
                 input("Error, resource file " + curFile + " not found.")
                 return
-
         edit_asset(curFile + "_JP", TL_path)
 
     input("All done! Press Enter to exit.\n")
 
-    # Okay, so, some of you are going to ask "why the fuck do we have to run a separate file."
-    # The answer is that unityPy, as far as I can tell, doesn't support closing the file handle without killing
-    # the entire process, and I'm not going to redistributed a modified version of the library because that
-    # would be hell.
-
 
 def edit_asset(inp, TL_path):
     env = UnityPy.load(inp)
-    env.files["globalgamemanagers.assets"] = env.load_file(os.path.join(src, "globalgamemanagers.assets"))
+    env.files["globalgamemanagers.assets"] = env.load_file(os.path.join(dataFolder, "globalgamemanagers.assets"))
+    print("Modifying file " + inp)
     # make sure that Texture2Ds are at the end
     objs = sorted(
         (obj for obj in env.objects if obj.type.name in TYPES),
@@ -163,7 +166,34 @@ def edit_asset(inp, TL_path):
         f.write(env.file.save())
 
 
+def csvToJsonData(TSVfileLocation):
+    data = {}
+    with open(TSVfileLocation, encoding="utf8") as f:
+        lines = f.readlines()
+        last = lines[-1]
+        lineIndex = 0
+        for line in lines:
+            lineIndex = lineIndex + 1
+            if lineIndex == 1:
+                try:
+                    # Check if the first cell is an integer.
+                    int(line[0:line.index("\t"):])
+                except ValueError:
+                    # It's not, assume these are the headers.
+                    continue
+            try:
+                currentID = line[0:line.index("\t"):]
+                tlText = line[line.rfind("\t") + 1:len(line):].rstrip()
+            except:
+                print("Error! Subtitles files is not formatted correctly, please check line " + str(lineIndex))
+                input("Press enter to exit.")
+                raise Exception("Incorrect formatting")
+            data[currentID] = tlText
+    return data
+
+
 def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
+    global MB_JSON
     if obj.type.name not in TYPES:
         return []
     if not STRUCTS:
@@ -171,8 +201,15 @@ def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
             STRUCTS.update(json.load(f))
 
     if not MB_JSON:
-        with open(os.path.join(TL_path, "subtitles.json"), "rt", encoding="utf8") as f:
-            MB_JSON.update(json.load(f))
+        JSONdata = csvToJsonData(os.path.join(TL_path, "subtitles.tsv"))
+        if os.path.isfile(os.path.join(TL_path, "subtitles.tsv")):
+            MB_JSON = csvToJsonData(os.path.join(TL_path, "subtitles.tsv"))
+        if os.path.isfile(os.path.join(TL_path, "subtitles.csv")):
+            MB_JSON = csvToJsonData(os.path.join(TL_path, "subtitles.csv"))
+        if not MB_JSON:
+            print("Error! Could not find subtitles tsv/csv file.")
+            input("Press enter to exit.")
+            raise Exception("Subtitles not found")
 
     data = obj.read()
 
@@ -185,12 +222,10 @@ def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
     if obj.type == ClassIDType.MonoBehaviour:
 
         # if TLDebug is set, modify all available monobehaviours
-
         if not str(obj.path_id) in MB_JSON:
-            if not str(obj.path_id) in DEBUG_DICT:
+            if not str(obj.path_id) in SPECIAL_OVR:
                 return [obj.path_id]
         # Only edit if the path id is in the dict
-
         # The data structure of MonoBehaviours is custom
         # and is stored as nodes
         if obj.serialized_type.nodes:
@@ -205,7 +240,7 @@ def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
 
             export = json.loads(export)
             obj.save_typetree(export)
-        else:  # I'm leaving this here even if for this game it shouldn't be necessary.
+        else:
             for i in [1]:
                 # abuse that a finished loop calls else
                 # while a broken one doesn't
@@ -228,56 +263,39 @@ def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
                 )
 
                 print("Patching MonoBehaviour with ID : " + str(obj.path_id))
+                try:
+                    currJSON = obj.read_typetree(nodes)
+                except Exception as e:
+                    print(e, obj.path_id, script.m_AssemblyName, script.m_ClassName)
+                    return [obj.path_id]
 
-                if str(obj.path_id) in DEBUG_DICT:
-                    # COMPLETELY DEBUG FEATURE, SHOULD NOT BE USED.
-                    # Replace the entire thing.
-                    with open(DEBUG_DICT[str(obj.path_id)], "rt", encoding="utf8") as f:
-                        export = f.read()
-                        export = json.loads(export)
-                        obj.save_typetree(export, nodes)
-                        break
+                # currJSON = obj.read_typetree(nodes)
+                if str(obj.path_id) in SPECIAL_OVR:
+                    print("Found a special override, apply it.")
+                    # Replace the entire JSON with the modified version.
+                    # This way we can still apply translations as normal.
+                    # Was a debug-only feature, now I use it for special cases so I don't have to hardcode them in.
+                    with open(SPECIAL_OVR[str(obj.path_id)], "rt", encoding="utf8") as f:
+                        currJSON = json.loads(f.read())
 
-                export = json.dumps(obj.read_typetree(nodes), indent=4, ensure_ascii=False)
-
-                # Check if this is textEffect, if so, disable it and break.
-                if obj.path_id == 226546:
-                    export = export.replace("\"m_Enabled\": 1,", "\"m_Enabled\": 0,")
-                    export = json.loads(export)
-                    obj.save_typetree(export, nodes)
-                    break
-
-                i = export.rfind("m_Text\": ") + len("m_Text\": ")
-                if export.rfind("m_Text\": ") == -1:
-                    break
-                beforeText = export[0:i + 1]
-                JPtext = export[i + 1:export.index("\n", i) - 1]
-                afterText = export[export.index("\n", i) - 1:len(export)]
-
-                # Check if this is the startup red text. If so, modify the line height.
-                if obj.path_id == 232224:
-                    beforeText = beforeText.replace("\"m_LineSpacing\": 1.0", "\"m_LineSpacing\": 1.1")
-
-                if MB_JSON[str(obj.path_id)] == "":
+                textItem = ""
+                if "m_text" in currJSON:
+                    textItem = "m_text"
+                elif "m_Text" in currJSON:
+                    textItem = "m_Text"
+                if textItem != "":
+                    newText = ""
                     if tlDebug:
-                        export = beforeText + "#" + str(obj.path_id) + " " + JPtext + afterText
-                        # Text is missing TL, just add the ID. More of an edge case than anything.
-                else:
-                    decoded = str(MB_JSON[str(obj.path_id)].replace("\n", "\\n").replace("\r", "\\r")) \
-                        .replace("\"", "\\\"")
-                    decoded = re.sub(r'(?<!\\)\\(?!["\\/bfnrt]|u[0-9a-fA-F]{4})', r'', decoded)
-                    if tlDebug:
-                        # Stupid encoding jank, I know it's ugly
-                        export = beforeText + "#" + str(obj.path_id) + " " + decoded + afterText
-                        # Add TL and ID.
+                        newText = "#" + str(obj.path_id) + " "
+                    if MB_JSON[str(obj.path_id)] == "":
+                        newText = newText + currJSON[textItem]
                     else:
-                        export = beforeText + decoded + afterText
-                        # Just add the TL.
-                # new_data = json.load(open("edited.json", "rt", encoding="utf8"))
-                export = json.loads(export)
-                obj.save_typetree(export, nodes)
-                break
+                        newText = newText + MB_JSON[str(obj.path_id)]
+                    # This reverse escapes special characters.
+                    currJSON[textItem] = newText.encode('raw_unicode_escape').decode('unicode_escape')
 
+                obj.save_typetree(currJSON, nodes)
+                break
     elif obj.type == ClassIDType.Texture2D:
         from PIL import Image
         if str(obj.path_id) in T2D_JSON:  # Check if ID is in JSON.
@@ -286,30 +304,9 @@ def overwrite_obj(obj, TL_path: str, append_name: bool = False) -> list:
             data.set_image(img_data, mipmap_count=data.m_MipCount)
             data.save()
 
-    elif obj.type == ClassIDType.Sprite:
-        if "idou" == data.name:
-            return [obj.path_id]
-
-            # This was a test to see if I could increase the size of the "MOVE" text.
-            # It did not work.
-
-            from PIL import Image
-            print(str(obj.path_id))
-            print(data.name)
-#            data.image = Image.open(os.path.join(TL_path, "textureOverrides", "idou #492.png"))
-            print(data.m_Rect.size)
-            data.m_RD.textureRect.bottom = 46
-            data.m_RD.textureRect.height = 46
-            data.m_RD.textureRect.size = (94, 46)
-            data.m_RD.textureRect.width = 94
-            data.m_RD.uvTransform.X = 100
-            data.m_RD.uvTransform.Z = 100
-            data.m_RD.uvTransform.W = 11.5
-            data.m_RD.uvTransform.Y = 23.5
-            print("")
-            data.save()
-
-
+    # elif obj.type == ClassIDType.Sprite:
+    #    if "idou" == data.name:
+    #       return [obj.path_id]
     return [obj.path_id]
